add docket-compose.yml file to postgree integration + change some attributes in User/Payroll/Employee class + Create API (POST) to create Users.

+++ src/main/java/io/github/progmodular/gestaorh/model/entities/DashboardResumo.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class DashboardResumo {
     private int totalFuncionarios;
     private int totalFolhasGeradas;
-    private double totalPagoNoMes;
+    private BigDecimal totalPagoNoMes;
 }
